Use local locks where possible for local temp tables.
Local temp tables should not need to access shared lock structures, so
reduce the lock level used to AccessShareLock - this makes the locks
eligible for fast-path locking, which is completely session-local.

TODO: Currently fast-path locking is limited to a maximum of 16 locks
per backend. We need to investigate if this has an impact on local
locking, and if the limit can be increased.

Signed-off-by: Jason Teng <jasonten@amazon.com>