Fix engine crash in tsql_openjson_with_get_subjsonb (#406)
This change resolves two engine crashes in tsql_openjson_with_get_subjsonb. The first crash was related to the way that strict and lax modes were handled in openjson column paths. In the case where a user did not put a space between strict/lax and the column name, it would cause the column type to be left NULL, resulting in a NULL entry into strlen and a crash. This change handles strict/lax modes properly, and adds a null check for column type to prevent crashes. The second crash was caused by a null value being passed into a json result list. In cases where a given json path did not exist, we would represent this by create a list of one null element. This element would be passed into the JsonbValueToJsonb function and result in crash. This commit removes the null entry from the list to prevent this crash from happening.

Task: BABEL-3820
Signed-off-by: Jake Owen <owjco@amazon.com>