reset proconfig for functions when babelfish is installed (#382)
If babelfish is installed the set_sql_dialect hook is installed. That means we are never resetting the proconfig values for functions with proconfig.
One major symptom of this is mVU failure when there are other extensions installed which executes a trigger which changes search path as part of proconfig.

Task: BABEL-4982
Signed-off-by: Tanzeel Khan <tzlkhan@amazon.com>